Changing Reader attributes related to sample frames
Instead of sample_frames_read and sample_frames_remaining,
changing to current_sample_frame and total_sample_frames.

This seems more future proof if the ability to seek to arbitrary
sample frames (instead of reading all samples consecutively) is
added in the future. (As in, sample_frames_read and
sample_frames_remaining don't necessarily make total sense if
that is possible).

Also, querying the total sample frames seems like a more common
operation than querying the sample frames remaining, so this
has the benefit of making getting that more direct.