Correcting examples and removing possible point of confusion
They wrote a half second of sound, but a comment indicated they
wrote 1 second.

To write a full second would require 441 writes of the cycle,
which is coincidentally similar to the sound's frequency of
440Hz. So, changing the number of writes to 100 to hopefully
be more clear that that number is unrelated to the frequency.
(This results in a short beep rather than a full second of sound).