Making buffer conversion faster in certain cases
Specifically, it takes the number of channels in the new and old
formats into account, in order to perform the bits-per-sample
conversion on the fewest number of samples.

For example, when converting stereo to mono, performing the bits
per sample conversion AFTER converting to mono will result in the
conversion operating on fewer samples. When converting mono to
stereo, performing the bits per sample conversion BEFORE converting
to stereo will also result in operating on fewer samples.