Not raising exception on Reader.new if unsupported Wave format
Allows querying metadata about valid but unsupported Wave files,
without having to have the info() method. Semantics are now that
a Reader can successfully be constructed to read a valid Wave file,
but read() will raise an error if the file is not supported by
this gem.

Adding Reader.native_format and Reader.readable_format? methods
as well. readable_format? indicates whether a call to read() will
be successful or not.