Extracting core Reader functionality to DataChunkReader
I.e., the core functionality to read samples from a data
chunk, and convert them into Buffers of the correct format,
etc.

The benefit of doing this is that it makes it easier to
support data chunk readers for other formats in the future,
such as compressed formats, AIFF, etc.