Adding UnvalidatedFormat.to_validated_format
Adding this allows logic to be moved from DataChunkReader to here,
which seems like a better place. The logic is more easily testable
here, and allows DataChunkReader to be more focused on just reading
the data from the file, rather than knowing how to convert the raw
format data in the file to a normalized format.