Chunk readers no longer need to handle padding byte
I.e., chunk readers are given the size of the actual amount of data
in the chunk, instead of the actual amount of data + padding byte.
This simplifies chunk reader implementations and can prevent
subtle errors from happening if an individual chunk reader
implementation doesn't handle the padding byte properly.