Removing Reader.file_name field
In advanced of allowing a Reader to be constructed with an IO,
in addition to a file name String. When constructed with an IO,
there won't necessarily be a sensible value for the file_name
field, and I'm not sure of a direct benefit of even having
this field in the first place, so just removing it altogether.