Removing Writer.file_name
For the same reason that Reader.file_name was removed in
a recent commit. When constructed with an IO, there won't
necessarily be a sensible value for the file_name field,
and I'm not sure of a direct benefit of even having
this field in the first place, so just removing it altogether.