Writer can be constructed with either a file name, or an open IO
Adding a smoke test for this, but it should be more exhaustively
tested in a future commit.

As part of this, changing Writer to use write/seek internally
instead of syswrite/sysseek. The reason is that StringIO doesn't
respond to sysseek, so is otherwise not usable with Writer.

The immediate driver for wanting to use StringIO is for tests,
but it also seems like it would be better for Writer to support
a wider range of IO-like objects such as StringIO in general.

My understanding is that the main difference between write and
syswrite is that write writes the data to an intermediate Ruby
buffer, to be flushed at a later time, while syswrite writes
the data immediately without buffering. I'm not sure what the
actual practical impact of that is, but it's not immediately
apparent to me whether it makes a difference or not for the
purposes here.

So, the tradeoff I'm making is in favor of the known benefit
(allowing use of StringIO), rather than the possible but
not apparent risk of using write instead of syswrite.