Adding tool for creating fixture Wave files
This is intended as a flexible tool for creating test fixture
Wave files. The ultimate goal is to eventually be able to use
it to automate the creation of all of the test fixture files
used for testing this gem.

The tool operates off of YAML files which determine which data
to write to each file via configuration. A few example YAML
files are included in this commit. In the future, the
creation of each fixture could be automated, for example
via a Rake task.

It's intentionally possible to create invalid Wave files.
This is to allowing creating test fixtures for invalid files.

To support all of the current test fixtures, the YAML format
and the tool will need to be made more flexible and given
more configuration options.