IO used to construct Reader no longer closed when Reader is closed
If a Reader is constructed from a file name, the underlaying
File opened internally by the Reader will still be closed when
closed? is called.

The reason for this behavior is that if the caller is passing in
an open IO instance to the constructor, they don't necessarily
want it to be closed when the Reader is closed. For example,
imaging reading multiple consecutive Wave files from a single
IO stream - you would want to keep the IO open between closing
and opening each Reader instance.