Seeking to end of IO when closing Writer
If a Writer instance is constructed from an IO (as opposed
to a String representing a file name), the byte offset
is now set to the end of the IO when the Writer is
closed, instead of the end of the header. This seems like
a more sensible byte offset - for example, it allows
writing consecutive files to the same IO without
overwriting the previous file.

This is not done when the Writer is constructed from a
file name, since the underlaying IO is also closed when
the Writer is closed.

Also extracting separate test methods for the writer-from-IO
and writer-from-file-name cases, since there are more diffences
between the two tests.