Correcting WAVE_FORMAT_EXTENSIBLE fixtures to use correct audio format
Previously, these fixture *.wav files were written to use either format
code 1 or 3, instead of the 65534 format code. This meant that the extra
WAVE_FORMAT_EXTENSIBLE data was written to these files, but the files had
the incorrect audio code.

This only affected tests, and not the actual gem implementation. This did
not cause test failures in the Reader tests, because the extra EXTENSIBLE
fields aren't actually exposed in the gem, so there were no tests to fail.
It just meant the files were silently read as non-EXTENSIBLE files.

This did not cause tests failures in the Writer tests, because the EXTENSIBLE
fixtures are ignored, because it's not currently possible to write files using
the EXTENSIBLE format.