Initial work to write relevant formats as WAVE_FORMAT_EXTENSIBLE
In this commit, a file is written a WAVE_FORMAT_EXTENSIBLE if:

* It's PCM and has more than 2 channels
* Or, it's PCM and has something other than 8 or 16 bits per sample

Although writing this type of file as a "vanilla" file seems to result
in files that play correctly in common media players, it's more correct
according to the spec to write these files in WAVE_FORMAT_EXTENSIBLE
format.

See http://msdn.microsoft.com/en-us/library/windows/hardware/gg463006.aspx

The actual implementation in this commit is messy, but passes the tests.
More work needs to be done to support writing non-PCM formats in EXTENSIBLE
format, at which time the implementation can be cleaned up.

Currently, a default value is always written to the speaker->channel mapping
field, regardless of what value is actually set on the `Format` instance.