Same number of speaker mappings as number of channels
If a `Format` is constructed and there are not enough speaker
mappings defined for the number of channels, the speaker mappings
array will be right-padded with :undefined. This allows for a
more consistent model in which there is always a speaker mapping
value for each channel (even if the mapping is undefined).

For example, when constructing this `Format` instance:

Format.new(3, :pcm_16, 44100, speaker_mapping: [:left_front])

...the speaker mapping will be converted to:

[:left_front, :undefined, :undefined]