Extra channel->speaker mappings removed when constructing `Format`
For example, if a `Format` specifies 2 channels, but speaker mappings
for 4 channels are given, the final 2 mappings will be removed when
constructing the instance.

The reason for doing this is to preserve the property of the
speaker mapping field to always have the same number of elements as
the number of channels.

A consequence of this is that you can pass in whatever you want as
additional elements in the speaker mapping array, and they will be
silently removed without causing an error. This seems like a less
fussy approach than raising an error, although maybe it's possible
raising an error/warning could alert the user to potential mistakes.
Going with the 'remove extra elements' for the time being, since on
the balance it seems like a nicer experience.