`UnvalidatedFormat` accurately reflects literal speaker mappings set
For example, if an `UnvalidatedFormat` is constructed with 3 channels
but a speaker mapping bit field that includes 5 speakers, then the 5
speakers will be preserved, instead of truncating the final 2.

This makes `UnvalidatedFormat` different from `Format`. `UnvalidatedFormat`
is intended to literally reflect what is defined in a source Wave file
being read, whether it is valid or not, while `Format` only allows
values that are sanitized and validated.

A benefit of this distinction is that the actual values defined inside
a source Wave file can be inspected, regardless of whether they are
valid or not.