`Format.channels`/`Format.sample_rate` can't be constructed with float
Previously, you could construct an instance with channels == 2.5 or
sample rate == 44100.7, and these values don't make sense. You would
probably get an error somewhere downstream if you tried to use these
values, but it's better to not allow them in the first place.

This implementation also rejects floats that are equivalent to an
integer, such as 2.0. While there is arguably a case for allowing
these, I'm not sure of a practical reason why this is necessary.