Better error message if `Format` constructed with invalid format code
The new error more directly states what the invalid code was, and what
the valid options are. In addition to being a better error message, the
validation implementation is simpler.

The `Format::SUPPORTED_BITS_PER_SAMPLE` constant is now only used by
tests and is a candidate to be removed, once the relavant test code is
refactored.