Files written as EXTENSIBLE if 1 or 2 channels, non-default speakers
If a file has 1 channel and 8 or 16 bits per sample, it will be written
using WAVE_FORMAT_EXTENSIBLE format if the speaker mapping is anything
other than [:front_center].

If a file has 2 channels and 8 or 16 bits per sample, it will be written
using WAVE_FORMAT_EXTENSIBLE format if the speaker mapping is anything
other than [:front_left, :front_right].

This _does_ mean that it's not possible to write a file using
WAVE_FORMAT_EXTENSIBLE format for these formats:

1 channel, 8 bits per sample, [:front_center] speakers
1 channel, 16 bits per sample, [:front_center] speakers
2 channels, 8 bits per sample, [:front_left, :front_right] speakers
2 channels, 16 bits per sample, [:front_left, :front_right] speakers

Instead, these formats will be written as a "vanilla" file with no
speaker mappings defined. In practice, I'm not sure why there would be
a reason you'd explicitly want a WAVE_FORMAT_EXTENSIBLE file for these
formats?