Fixing tests that were inadvertantly not testing anything
The previous tests for writing to a StringIO instance weren't
actually testing anything! The data written to the StringIO wasn't
actually looked at. This is now fixed.

Fortunately, with a correct test implementation none of the tests
fail, so it doesn't seem like the 'write to IO' functionality was
inadvertantly broken.