Tests for `Writer` properly clean up generated test files
I.e., the `test/fixtures/actual_output` folder is now empty when
the `Writer` tests are complete. Previously, the most recent
generated file would still be around.

The previous behavior had the benefit of the offending file being
around for inspection if a test failed. However, in the future the
can be replicated by commenting out the `clean_output_folder` call
in `teardown`.