Extra data in `inst` chunk will be read correctly
This is a result of testing with a file that contains an 'inst'
chunk that has a size of 8, instead of the expected 7. (Presumably
to make the chunk an even size, but who knows).

It seems like this should maybe be a general purpose thing added
to all chunk readers, but doing it just for the 'inst' chunk for now.