File header written in correct location when writing to arbitrary IO
Previous, the header was always written starting at byte 0 of the stream.
However, it should really be written at the seek position the stream was
at when writing the file began.

For example, this now allows writing consecutive files to the same stream.
Previously, the first file would have had a corrupt header (because it
would be overwritten with incorrect data from the subsequent files), and
the additional files would have had incomplete headers that indicated a
sample count of 0.