Simpler way of determining if at end of RIFF chunk
This no longer requires counting the number of bytes that have
been read so far, which is more complicated, error prone, and
subject to issues if a chunk's stated size doesn't match it's
actual size.

This change exposed bad data in one of the fixtures, so correcting
that fixture as well.