Removing extraneous paragraph
This comment is no longer 100% necessary since `each_buffer` uses
a default value for the sample frame count if one isn't given. It
doesn't directly make sense with the example as-is, because the
example doesn't use a custom buffer size.

The removed info is still mentioned in the rdoc for `each_buffer`,
where is seems more relevant.