Reducing seeking to help IO instances that don't support seeking
Ignored chunks are now skipped by reading their data, instead of
seeking past them. This allows for better compatibility with IO
instances that don't support seeking (such as sockets). It should
allow `Reader` to work properly with this type of IO on files that
don't contain any chunks after the data chunk. Errors could still
be encountered if there are chunks after the data chunk, because
seeking will still be required in that case.