Fixing allowed range for `SamplerLoop.fraction`
The spec indicates that 0x8000_0000 represents 0.5, which would
imply that 1.0 is 0x1_0000_0000. That would make the maximum
value 0xffff_ffff equivalent to 0.999999999767169. Rather than
checking for that very exact float value, instead checking that
the value is less than 1.0, and assuming that a hypothetical
future ability to write a `smpl` chunk will quantize values
between 0.999999999767169 and 1.0 down to 0.999999999767169.
(Which seems like a situation very unlikely to actually be
encountered, but at least it is more thoeretically correct).