Adding validation that `SamplerInfo` fields are in allowed range
This doesn't test that all values are in the semantically correct
range. For example, it doesn't check that the MIDI note is between
0 and 127. However, it validates that all values should be within
the ranges that can be represented in an output *.wav file. For
example, if a field is stored as a 32-bit unsigned integer, the only
possible values are between 0 and 4_294_967_295.

The benefit is that this should remove a class of error caused by
constructing a `SamplerInfo` object with values that can't actually
be written to a file (without at least some conversion to a different
value). Making sure the values make sense is still left up to the
user.

This doesn't validate the `smpte_offset` field yet, because that
field might undergo some changes.