Attempting to remove misleading variable value
Previously, `data_chunk_is_last_chunk` could be equal to false
even though the data chunk was the final chunk. This could occur
if the data chunk had an odd size and required a padding byte.
This made it a little confusing for me when trying to trace the
code.

I'm not sure if the new version is an improvement, because it
requires an extra place where the padding byte is taken into
account. But, at least it means that variable values should no
longer be potentially misleading?