Improving fixture *.wav files to be generated from YAML files
Previously, these two fixture files could not be easily changed
because they weren't generated from a corresponding YAML file. The
blocker was that the YAML format didn't support a way to add padding
bytes, but it now does (in a sort of makeshift way by using an
"extra_bytes" field in the data chunk configuration).