Fixing debug tool to output data in "adtl" sections correctly
Previously, the sub chunks in an "adtl" list section were treated
the same as sub chunks in an "INFO" list. That is, as raw strings.
However, they actually consist of two fields: a cue point ID, followed
by a raw string.

This also makes a change to interpret cue point IDs as integers,
instead of strings.