Normalizing a sampler loop play count of 0 to Float::INFINITY
A value of 0 in this field means to play the loop indefinitely,
instead of a fixed number of times.

I went back and forth on whether to do this normalization, but
ultimately decided to do it. Ultimately "infinity" _is_ the
semantically correct value, and "0" is an artifact of how the
bits are stored on disk. Although someone well versed with the
file format might be surprised to find a value of "infinity"
instead of "0", it doesn't seem like it would be confusing?
Whereas it seems more obviously the case how blindly returning
0 and expecting the user to know that that doesn't actually mean
0 would be confusing.