Incomplete data chunk size field no longer considered valid
A file that contains a data chunk with a size of 0 is valid. However,
this is only the case if the data chunk size field contains a full 4
bytes all set to 0. Previously, if the chunk size field was fewer than
4 bytes followed by the end of the file, the file would be considered
valid even though the chunk size field was incomplete.

In addition to handling this specific case, this change also avoids
normalizing an incomplete size field to 0, to make the logic easier
to understand and prevent other possible issues in the future.