Fixing bug where `Reader.current_sample_frame` was incorrect
If the data chunk listed a chunk size that was larger than the
actual chunk size (as determined by the actual bytes in the file),
attempting to read past the end of the chunk would incorrectly set
`Reader.current_sample_frame` to a value past the end of the chunk.

For example, suppose a file's data chunk has 1000 bytes, but the
chunk header indicates 2000 bytes. Attempting to read 1500 bytes
would previously correctly only return the available 1000 bytes,
but would set `Reader.current_sample_frame` to (1500 / block_align)
instead of (1000 / block_align). Now it will correctly be set to
(1000 / block_align).

This bug did not occur for files in which the data chunk listed
the correct size.