Adding test that final chunk can be missing padding byte
If the final chunk has an odd size but is missing the padding byte,
this will be considered OK as long as the RIFF chunk size matches
the actual number of bytes in the file. Not raising an error in this
scenario is intended as a pragmatic choice to be more flexible in
which files can be successfully read, while also being simpler to
implement.