Inlining class with a method
There doesn't seem to be a particular reason for this method to
have its own class, and the previous implementation was a little
confusing because the class was called `RiffChunkReader`, even though
it only read the format code but not the rest of the chunk (i.e. the
child chunks). `RiffReader` is now essentially the full "riff chunk
reader" class.