Simulating required keyword arguments for Ruby 2.0
Ruby 2.0 doesn't support the syntax for required keyword arguments,
i.e. `my_required_arg1:, my_required_arg2:`. This syntax will raise
an error. Either dropping support for 2.0 just for this reason, or
making these keywords not required (which would cause backwards
compatibility issues if done in a future version) aren't appealing.

The error message that one receives in the case of a missing keyword
will not necessarily be the same as Ruby 2.1+, but the exception
class will consistently be `ArgumentError`.