Using correct way of making constants private
I didn't realize that placing constants after `private` isn't the
correct way to create a private constant. Since `private` is apparent
not a keyword but a method, defining constants in this way is
seemingly equivalent to defining a constant inside a method?

There are other places in the gem where private constants are
defined in this way, but in this comment only changing constants in
new classes added since v1.0.1.