Preventing warning when reading file with "smpl" chunk in Ruby 2.7
Apparently Ruby 3 will introduce changes in the way a hash passed as a
method argument is converted to keyword arguments. Starting in Ruby 2.7,
certain code which will break in Ruby 3 will cause a warning to be emitted.
Due to the way `SampleChunkReader` constructed new `SamplerInfo` and
`SamplerLoop` instances internally, this caused a "warning: Using the last
argument as keyword parameters is deprecated; maybe ** should be added to
the call" error to be emitted when reading a file with this type of chunk
using Ruby 2.7.

This commit removes the warning when using Ruby 2.7, and should make the
code future-compatible with Ruby 3. (At least, according to
https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/).