Fixing bottom margin of "List of Features" section on home page
When the home page was viewed in a desktop-sized viewport, the "List
of Features" section had extra margin at the bottom, due to the bottom
margin of the final row in the list of features. This was previously
handled specially for the mobile viewport case, but not the desktop
viewport case.