Preventing content overflow in just-wider-than-mobile viewport size
Previously, when the viewport was just barely wider than the mobile
viewport size, some content expanded out of the page, causing the
page to be horizontally scrollable. In particular, this affected the
site when viewing on an iPad in portrait orientation.

In addition, this improves the appearance at this viewport size by
adding additional page padding to give the content room to breathe.