Using absolute URLs, instead of relative
This allows the main pages to be visited via a URL ending with a `/`
(e.g. `/examples` previously worked but now `/examples/` does too),
and fixes the 404 page to not be unstyled when going to certain URLS
(such as any ending with an `/`) due to the main `wavefile.css` file
not loading correctly.