Simplifying CSS implementation of margins for <pre> tags
The margins for <pre> tags are now defined in a single place in the
CSS (per breakpoint), instead of in two places.

As a side benefit, this also allows removing the setting of default
`auto` left/right margins for each child of a `.section` container.
(I.e. this is a benefit because the fewer styles that are set on
elements en masse, the easier the CSS is to understand/work with).