Changing layout to fill screen on iPhone X-style phones
In practice, this means when in portrait mode the background color
of each section should extend to reach the edges of the screen,
instead of only reaching the safe inset area.

Hopefully this commit works, since this hasn't been tested on an
iPhone X-style phone! (We'll see after deploying this to production,
i.e. GitHub Pages).