Adding syntax highlighting to code block
This code block was previously not using the normal syntax
highlighter because I wanted to use a particular CSS style to
emphasize part of the code, but I don't know how to escape markup
inside a syntax highlighted block so that it is interpreted as
HTML, and not text to be highlighted.

Instead, this commit opts for using a code comment to highlight
the relevant changes, so that syntax highlighting can be used.