Debug script properly handles extra data in format chunks
Previously, if a format chunk contained extra trailing bytes at
the end, the extra bytes would not be properly handled. This would
cause the reading position to become offset from the correct
position, and subsequent chunks would not be read properly.

This also prevents attempts to interpret fields that don't exist
for the given format code. For example, if the format code is 1,
then any data after the "bits per sample" field will be displayed
in an "extra bytes" section, instead of e.g. being shown as an
"extension size" field.