Adding test for format chunk with missing extension size
Normally, if the format code in the format chunk is not "1", then
the format chunk must contain an "extension size" field. If the
format code is "3", then this field should be set to "0", and this
should be the final field in the format chunk.

However, files with a format code of "3" that are missing the
"extension size" field can still be read by the gem, as long as
the chunk size reflects the missing field. There isn't actually
any information lost by the absence of this field, and this makes
the gem a little more flexible in terms of which files it can
read.

Previously there was no explicit test case for this scenario, so
this commit adds one. This is being done in advance of make some
changes to the logic for reading the format chunk, to reduce the
chance of an accidental regression.