Format chunks with extra trailing bytes are read properly
Previously, if a file's format chunk contained extra trailing
bytes at the end, an `InvalidFormatException` was sometimes raised
when the file was read. (If the extra bytes happened to matched up
with the format of a WAVE_FORMAT_EXTENSIBLE extension, then the
file could be successfully read, although the extension would be
ignored).

Now extra trailing bytes will be ignored in any format chunk,
regardless of the format. This seems like a more flexible
approach, and there doesn't seem to be any downside to ignoring
the extra data.