Better "Valid Bits Per Sample" field description
The old description was confusing because made it seem like the
only way to have a bits per sample value that was different from
the sample container size was to use this field, even though the
earlier "bytes per sample frame" description shows a different way
to do this. It also was missing an explanation of the historical
reasons for why this field was added.