Error raised if extensible format chunk extension too short
Previously, a file in which the format chunk was in extensible
format (i.e. the format code was 65534) but the format chunk
extension size field reported a size that was too small (i.e. less
than 22 bytes) could be opened using `Reader.new()`. However, no
data could be read out of the file because `readable_format?` was
set to false, and `Reader.native_format` would have `nil` values
for the missing fields.

This seemed like undesireable behavior, because it seems
misleading to successfully allow creating a `Reader` instance for
an invalid file. There aren't reasonable default values that can
be assumed when fields are missing from the extension, so it seems
better to raise an error in this scenario.

This change requires a change to `tools/fixture_writer.rb` to
allow fixtures with an incomplete extensible format chunk
extension to be created.