Debug script reads extra bytes at end of `smpl` chunk
Previously, extra bytes at the end of a `smpl` chunk would not be
read, causing a premature attempt to start reading the next chunk.
This would result in the next chunk not being found correctly, and
prevent reading the rest of the file. Now these extra bytes will
be read and accounted for.