Debug script more helpfully indicates misaligned chunks
Previously, if an error occurred due to a misaligned chunk the
error would be displayed and the script would exit. However, the
error did not always make it clear that the problem was that the
next chunk was not found properly.

Now, before attempting to read an unrecognized chunk the attempted
chunk ID and chunk size will be displayed. In the case of an
error, this should make it more clear what the problem was. For
example if the final script output is something like "'act' chunk
of size 3221225472, skipping.", the invalid chunk ID and huge
chunk size should more clearly indicate that something caused
chunk reading to become misaligned than an "Invalid argument @
rb_io_sysread" error.