Clarifying potentially confusing wording on home page
It seemed potentially confusing to say that the gem could query
metadata from a file that is "in a format this gem can't read or
write". If a file is in a format that can't be read, then how can
the gem read the metadata?

The new wording hopefully makes it more clear that the metadata
can be read regardless of whether or not the _sample data_ is in
a format that can be read.