Adding tests for odd sized extra format chunk data
There were previously already tests that tested what happens when
a format chunk with format code 1 has extra data at the end, but
not specifically tests for when the extra data contains an odd
number of bytes. (Important due to the need to account for an
extra padding byte when the chunk size is odd).