Adding tests for format chunks with oversized extensions
Format chunk extensions seem similar to format chunks as a whole
in that it seems valid for them to contain more bytes than the
normal expected size, as long as those bytes are ignored when
reading the file. This functionality is already supported by the
existing code, but this commit adds some tests to verify that.