Extra chunk extension bytes accounted for
Previously, the debug script did not account for the possibility
of extra bytes at the end of a WAVE_FORMAT_EXTENSIBLE format chunk
extension. This resulted in a premature attempt to read the next
chunk, and the contents of the file not being displayed correctly.
(This did not happen for other types of format chunk extension).

Now these extra bytes will be accounted for and displayed.