Consolidating logic for handling extra chunk bytes
The debug script logic to handle reading extra bytes at the end of
a format chunk was essentially duplicated for the case of a chunk
with format code 1, and for the case of a chunk with a different
format code. Now this logic is consolidated, which also seems to
make the code a little easier to read.