Adding tests for incomplete extension size
Since the format code for these two files is `1`, the extra byte
at the end of the format chunk shouldn't be treated as part of the
extension size.

The valid file (i.e. the one with the padding byte) will cause an
error in v1.1.1, but this test should validate that that bug has
now been fixed. The invalid file (i.e. the one without the padding
byte) will raise `InvalidFormatError` in both v1.1.1 and the code
as of this commit, but the error message should be different. (In
v1.1.1, "Not a supported wave file. The format chunk is missing an
expected extension", and as of this commit, "Not a supported wave
file. Unexpected end of file.").