Adding tests for incomplete extension size
Since the format code for both of these files is `65534` (i.e.
WAVE_FORMAT_EXTENSIBLE), an extension is required. The fact that
one of the files has the required padding byte and the other
doesn't shouldn't cause any behavior to be different, because the
missing extension should cause an error to be raised before the
missing padding byte does.