Removing redundant source of error in fixture
Previously, there were two different reasons why this fixture
would cause `UnsupportedFormatError` to be raised when attempting
to read sample data from it: the audio format code, and the sample
rate. To make sure that tests are testing what they are intended
to test, the fixture is being changed so that only the sample rate
has an unsupported value.

Since the format code is being changed to `1`, there is no longer
a need for the format chunk extension size field, so that is being
removed as well.