More accurate format code for test fixture
A file with an audio format code of `2` would normally be expected
to have a non-empty format chunk extension. To make this fixture a
little more accurate, the format code is being changed to a
different unsupported format code that actually _should_ have an
empty extension. Alternately, the format code could be kept as `2`
and a full extension added, but the extension isn't the point of
this fixture so changing the format code seems simpler.

This shouldn't make any practical difference with the current
version of the gem since it never tries to parse the extension of
a format chunk with an unsupported format code (other than
checking to make sure it doesn't overflow the chunk), but it seems
like good hygiene to make this change.