Adding summary of v1.1.2 changes
The detailed description of the v1.1.2 changes is kind of a lot
and possibly makes it harder to see the forest for the trees. This
summary hopefully makes it easier for someone to see at a glance
whether the changes in this release are likely to affect them or
not.