More accurate description in release notes
The previous description incorrectly stated that the chunk
extension would always be larger than expected when the error
occurred, but the error was also sometimes raised when the chunk
extension was the expected size. The new description also
hopefully more clearly indicate how the error message was
inaccurate when the format code was `1`.