Better labeling of FourCC values
This changes the data type for chunk IDs to be displayed as
"FourCC" instead of the chunk ID itself. (And the same for the
one-off case of the RIFF chunk form type). Other places where the
type of a FourCC value was displayed as "alpha_4" are also changed
to "FourCC" as well. It seems helpful to more specifically label
FourCC values as such.

Displaying the expected chunk ID (and RIFF chunk form type) along
with the actual chunk ID (or RIFF chunk form type) didn't seem
that useful, since seemingly the only time they would ever differ
is if bytes 0-3 are not "RIFF", or bytes 8-11 are not "WAVE".