Better display format for byte sequences
Previously, the parsed value for an arbitrary byte sequence was
displayed as a string, and the individual bytes were shown as
individual characters. However, it seems confusing to display a
value that semantically is not a string as a string, and it seems
more useful to display each individual byte as a numeric value.

In service of that, the parsed value of a raw byte sequence is now
displayed as "N/A" (because it doesn't have a clear value, and a
blank value seems potentially ambiguous), and each individual byte
value is displayed as an integer, instead of a character.