Displaying all byte values as a number
Previously, the individual bytes in strings and string-like values
(e.g. a FourCC) were displayed a character, and in bit fields or
hexadecimal fields were displayed as custom string. However, it
seems more useful to display the raw byte values as a number,
because the individual characters seem redundant since they are
also shown in the parsed value.