Better display format for unrecognized chunk types
Previously, an unrecognized chunk would be displayed like this:

  'nope' chunk of size 1024, skipping.

Now a full chunk header will be displayed, so the individual bytes
can be inspected:

  "nope" Chunk (unrecognized chunk type)
  ================================================================
  Chunk ID:              FourCC     | nope       | 6e 6f 70 65
  Chunk size:            uint_32    | 1024       | 00 04 00 00
  ----------------------------------+------------+----------------
  (chunk body omitted)

This is particularly useful when a missing padding byte causes
reading to become misaligned, resulting in the parsed chunk ID and
size being gibberish.

And speaking of padding bytes, an additional benefit of this
change is that if a padding byte is present, it will now be more
clearly tied to the chunk in question. Before:

  'nope' chunk of size 1023, skipping.
  Padding Byte:          byte       | 0          | 00

And after:

  "nope" Chunk (unrecognized chunk type)
  ================================================================
  Chunk ID:              FourCC     | nope       | 6e 6f 70 65
  Chunk size:            uint_32    | 1023       | ff 03 00 00
  ----------------------------------+------------+----------------
  (chunk body omitted)
  Padding Byte:          byte       | 0          | 00