Removing duplicated "display chunk header" logic
This removes the need to duplicate the "display chunk header"
logic in each chunk reader method, and also removes the need to
pass the chunk ID FourCC as an argument to each method as well.
As a result of this commit, each chunk reader method is now only
responsible for reading/displaying the chunk body, and not the
chunk header.