Fixing erroneous attempt to read padding byte
44e0a40afe0fa40ee73a7a2db4cabe57791dc2af introduced a bug that
caused an attempt to read a padding byte after the "form type"
field at the start of the "RIFF" chunk, if the "RIFF" chunk had an
odd size. However, a padding byte will never be in this location,
and this caused the file reader to become misaligned with the
start of the first child chunk.

Conceptually, it doesn't really seem to make sense for a "RIFF"
chunk to have an odd size, because all of its child chunks should
have an even size (when taking padding bytes into consideration).
That is, adding even numbers together should result in an another
even number. However, there isn't anything stopping a "RIFF" chunk
from reporting an odd size, and this commit makes sure that
scenario is handled properly.